requirements.txt
pandas
pyodbc
Requests
SQLAlchemy
python-dotenv
azure-functions
